+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [5.68.0](https://github.com/informatievlaanderen/public-api/compare/v5.67.2...v5.68.0) (2023-03-15)
+
+
+### Features
+
+* building grb upload ([935b925](https://github.com/informatievlaanderen/public-api/commit/935b9252098c16d1d33e4af99bbe620a5bb34d0e))
+
+## [5.67.2](https://github.com/informatievlaanderen/public-api/compare/v5.67.1...v5.67.2) (2023-03-14)
+
+## [5.67.1](https://github.com/informatievlaanderen/public-api/compare/v5.67.0...v5.67.1) (2023-03-14)
+
+
+### Bug Fixes
+
+* documentation page ([e121a7a](https://github.com/informatievlaanderen/public-api/commit/e121a7a77dc71d4baf6cb742b683974a4ad34450))
+
+# [5.67.0](https://github.com/informatievlaanderen/public-api/compare/v5.66.2...v5.67.0) (2023-03-13)
+
+
+### Bug Fixes
+
+* AddressWasRemovedBecauseStreetNameWasRemoved tags ([ec913ac](https://github.com/informatievlaanderen/public-api/commit/ec913ac721c746ef8a33fe9c3abe9a835c5a9d6b))
+
+
+### Features
+
+* add address match v2 ([eeb3147](https://github.com/informatievlaanderen/public-api/commit/eeb31478fef45ddad0a988078750af8df4ef4652))
+
 ## [5.66.2](https://github.com/informatievlaanderen/public-api/compare/v5.66.1...v5.66.2) (2023-03-06)
 
 ## [5.66.1](https://github.com/informatievlaanderen/public-api/compare/v5.66.0...v5.66.1) (2023-03-06)